requirements.txt
pydantic-settings>=2.6.0
PyYAML>=6.0.2
anthropic>=0.84.0
openai>=2.28.0
google-genai>=1.67.0
